\name{TFMPvalue-package}
\alias{TFMPvalue-package}
\docType{package}
\title{
  Efficient and accurate P-value computation for Position Weight Matrices
}

\description{
  This package provides a novel algorithm that solves the P-value calculation
  problem given the score based on a Postion Weight Matrices (PWMs), 
  or the reverse problem: finding the score give the desired P-value.
  This package is an interface to code originally made available by 
  Helene Touzet and Jean-Stephane Varre, 2007, Algorithms Mol Biol:2, 15.
}

\details{
  The original code is taken from 
  \url{http://bioinfo.lifl.fr/TFM/TFMpvalue/TFM-Pvalue.tar.gz}, 
  retrived 26/03/2014.

  The algorithm is described in Touzet, H., and Varre, J.-S. (2007). 
  Efficient and accurate P-value computation for Position Weight Matrices. 
  Algorithms Mol Biol 2, 15.
}

\author{
  Ge Tan
}
\keyword{package}
